+++ realm/realm-library/src/main/cpp/objectserver_shared.hpp
@@ -33,6 +33,7 @@
 #include "jni_util/jni_utils.hpp"
 #include "jni_util/java_global_weak_ref.hpp"
 #include "jni_util/java_method.hpp"
+#include "jni_util/jni_exceptions.hpp"
 
 
 // Wrapper class for realm::Session. This allows us to manage the C++ session and callback lifecycle correctly.
@@ -82,8 +83,7 @@ public:
                     local_env->CallVoidMethod(
                             obj, notify_error_handler, error_code.value(), local_env->NewStringUTF(message.c_str()));
                     if (local_env->ExceptionCheck() == JNI_TRUE) {
-                        local_env->ExceptionDescribe();
-                        throw std::runtime_error("Exception in error handler");
+                        throw realm::jni_util::JniPendingException("Exception in error handler");
                     }
                 });
             }